comment refactoring
